requirements.txt
anthropic>=0.79.0
dissect>=3.20.1
flask>=3.1.2
httpx>=0.28.1
jinja2>=3.1.6
markupsafe>=3.0.3
openai>=2.21.0
pyyaml>=6.0.3
